+++ widgets/BluetoothPill.qml
@@ -91,6 +91,22 @@ Rectangle {
     Layout.preferredWidth: implicitWidth
     Layout.preferredHeight: implicitHeight
     Layout.alignment: Qt.AlignVCenter
+    clip: false
+    DockFace {
+        id: dockFx
+        bar: root.bar
+        host: root
+        hovered: root.btHovered
+    }
+    transform: [
+        Scale {
+            xScale: dockFx.mag
+            yScale: dockFx.mag
+            origin.x: root.width / 2
+            origin.y: dockFx.growUp ? root.height : 0
+        },
+        Translate { y: dockFx.jumpY }
+    ]
 
     HoverHandler {
         id: btHoverVis
@@ -817,6 +833,7 @@ Rectangle {
         }
 
         onClicked: (mouse) => {
+            dockFx.jump()
             if (mouse.button === Qt.RightButton) {
                 bt.togglePower()
                 return